+++ src/main/java/org/sysc4907/votingsystem/controllers/AuthenticationController.java
@@ -1,4 +1,4 @@
-package org.sysc4907.votingsystem;
+package org.sysc4907.votingsystem.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,6 +6,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.sysc4907.votingsystem.services.AuthenticationService;
 
 @Controller
 public class AuthenticationController {
