+++ src/main/java/org/sysc4907/votingsystem/notinuse/VoterRepository.java
@@ -1,7 +1,7 @@
 package org.sysc4907.votingsystem.notinuse;
 
 import org.springframework.data.repository.CrudRepository;
-import org.sysc4907.votingsystem.VoterAccount;
+import org.sysc4907.votingsystem.models.VoterAccount;
 
 import java.util.Optional;
