+++ src/main/resources/templates/successful-voter-login.html
@@ -40,7 +40,7 @@
 <div class="container">
     <h1 th:text="'Welcome VOTER to our secure & transparent voting system, ' + ${name} + '!'"></h1>
     <div class="footer">
-        <p>&copy; 2024 Your Voting System</p>
+        <p>&copy; 2024 Secure & Transparent Voting System</p>
     </div>
 </div>
 </body>
