+++ src/test/java/org/sysc4907/votingsystem/VotingSystemApplicationTests.java
@@ -8,6 +8,7 @@ class VotingSystemApplicationTests {
 
 	@Test
 	void contextLoads() {
+
 	}
 
 }
